Fix sidebar position
