+++ packages/ui/src/components/Designer/RightSidebar/DetailView/index.tsx
