+++ packages/ui/src/components/Designer/RightSidebar/ListView/index.tsx
@@ -1,6 +1,6 @@
 import React, { useContext, useState } from 'react';
 import type { SidebarProps } from '../../../../types';
-import { SIDEBAR_WIDTH } from '../../../../constants';
+import { RIGHT_SIDEBAR_WIDTH } from '../../../../constants';
 import { I18nContext } from '../../../../contexts';
 import { getSidebarContentHeight } from '../../../../helper';
 import { theme, Input, Typography, Divider, Button } from 'antd';
@@ -76,7 +76,7 @@ const ListView = (
             style={{
               paddingLeft: 30,
               height: height - headHeight,
-              width: SIDEBAR_WIDTH - 35,
+              width: RIGHT_SIDEBAR_WIDTH - 35,
               lineHeight: '2.75rem',
             }}
           />
